Relocates the theme manager outside the app package
