+++ tests/test_ui.py
@@ -137,13 +137,13 @@ class TestMainWindow(QTestCase):
 
     def testThemeToggle(self):
         # Verifies the theme is toggled by the toolbar action
-        light_expected = config.load_stylesheet('style.css')\
-            if platform.system() != 'Darwin' else config.load_stylesheet('mac_style.css')
+        light_expected = self.window.themes.loadStylesheet('style.css')\
+            if platform.system() != 'Darwin' else config.loadStylesheet('mac_style.css')
         light_theme = self.window.styleSheet()
         self.assertEqual(light_theme, light_expected)
         QTest.mouseClick(self.toolbar.widgetForAction(self.window.theme_action), Qt.MouseButton.LeftButton)
         dark_theme = self.window.styleSheet()
-        dark_expected = config.load_stylesheet('dark_theme.css')
+        dark_expected = self.window.themes.loadStylesheet('dark_theme.css')
         self.assertEqual(dark_theme, dark_expected)
 
     def createProject(self):
@@ -852,70 +852,70 @@ class TestMainWindow(QTestCase):
         config.setup_logging(log_filename)
         self.assertTrue((config.LOG_PATH / log_filename).exists())
 
-        self.assertTrue(config.settings.value(config.Key.Align_First))
-        config.settings.setValue(config.Key.Align_First, False, True)
-        self.assertFalse(config.settings.value(config.Key.Align_First))
-        config.settings.setValue(config.Key.Align_First, "true", True)
-        self.assertTrue(config.settings.value(config.Key.Align_First))
-        config.settings.setValue(config.Key.Align_First, -2, True)
-        self.assertTrue(config.settings.value(config.Key.Align_First))
+        self.assertTrue(config.settings.value(config.settings.Key.Align_First))
+        config.settings.setValue(config.settings.Key.Align_First, False, True)
+        self.assertFalse(config.settings.value(config.settings.Key.Align_First))
+        config.settings.setValue(config.settings.Key.Align_First, "true", True)
+        self.assertTrue(config.settings.value(config.settings.Key.Align_First))
+        config.settings.setValue(config.settings.Key.Align_First, -2, True)
+        self.assertTrue(config.settings.value(config.settings.Key.Align_First))
 
-        item = config.__defaults__[config.Key.Local_Max_Eval]
-        self.assertEqual(config.settings.value(config.Key.Local_Max_Eval), item.default)
-        config.settings.setValue(config.Key.Local_Max_Eval, item.limits[1] + 1, True)
-        self.assertEqual(config.settings.value(config.Key.Local_Max_Eval), item.default)
-        config.settings.setValue(config.Key.Local_Max_Eval, item.limits[0] - 1, True)
-        self.assertEqual(config.settings.value(config.Key.Local_Max_Eval), item.default)
-        config.settings.setValue(config.Key.Local_Max_Eval, item.limits[1] - 1, True)
-        self.assertEqual(config.settings.value(config.Key.Local_Max_Eval), item.limits[1] - 1)
+        item = config.settings.default(config.settings.Key.Local_Max_Eval)
+        self.assertEqual(config.settings.value(config.settings.Key.Local_Max_Eval), item.default)
+        config.settings.setValue(config.settings.Key.Local_Max_Eval, item.limits[1] + 1, True)
+        self.assertEqual(config.settings.value(config.settings.Key.Local_Max_Eval), item.default)
+        config.settings.setValue(config.settings.Key.Local_Max_Eval, item.limits[0] - 1, True)
+        self.assertEqual(config.settings.value(config.settings.Key.Local_Max_Eval), item.default)
+        config.settings.setValue(config.settings.Key.Local_Max_Eval, item.limits[1] - 1, True)
+        self.assertEqual(config.settings.value(config.settings.Key.Local_Max_Eval), item.limits[1] - 1)
 
-        item = config.__defaults__[config.Key.Angular_Stop_Val]
-        self.assertEqual(config.settings.value(config.Key.Angular_Stop_Val), item.default)
-        config.settings.setValue(config.Key.Angular_Stop_Val, item.limits[1] + 1, True)
-        self.assertEqual(config.settings.value(config.Key.Angular_Stop_Val), item.default)
-        config.settings.setValue(config.Key.Angular_Stop_Val, item.limits[0] - 1, True)
-        self.assertEqual(config.settings.value(config.Key.Angular_Stop_Val), item.default)
-        config.settings.setValue(config.Key.Angular_Stop_Val, item.limits[1] - 1, True)
-        self.assertEqual(config.settings.value(config.Key.Angular_Stop_Val), item.limits[1] - 1)
+        item = config.settings.default(config.settings.Key.Angular_Stop_Val)
+        self.assertEqual(config.settings.value(config.settings.Key.Angular_Stop_Val), item.default)
+        config.settings.setValue(config.settings.Key.Angular_Stop_Val, item.limits[1] + 1, True)
+        self.assertEqual(config.settings.value(config.settings.Key.Angular_Stop_Val), item.default)
+        config.settings.setValue(config.settings.Key.Angular_Stop_Val, item.limits[0] - 1, True)
+        self.assertEqual(config.settings.value(config.settings.Key.Angular_Stop_Val), item.default)
+        config.settings.setValue(config.settings.Key.Angular_Stop_Val, item.limits[1] - 1, True)
+        self.assertEqual(config.settings.value(config.settings.Key.Angular_Stop_Val), item.limits[1] - 1)
 
-        item = config.__defaults__[config.Key.Fiducial_Colour]
-        self.assertEqual(config.settings.value(config.Key.Fiducial_Colour), item.default)
-        config.settings.setValue(config.Key.Fiducial_Colour, (2, 3, 4, 5), True)
-        self.assertEqual(config.settings.value(config.Key.Fiducial_Colour), item.default)
-        config.settings.setValue(config.Key.Fiducial_Colour, (2, 3, 4), True)
-        self.assertEqual(config.settings.value(config.Key.Fiducial_Colour), item.default)
-        config.settings.setValue(config.Key.Fiducial_Colour, ("h", "1.0", "1.0", "1.0"), True)
-        self.assertEqual(config.settings.value(config.Key.Fiducial_Colour), item.default)
-        config.settings.setValue(config.Key.Fiducial_Colour, ("1.0", "1.0", "1.0", "1.0"), True)
-        self.assertEqual(config.settings.value(config.Key.Fiducial_Colour), (1, 1, 1, 1))
-        config.settings.setValue(config.Key.Fiducial_Colour, (2, 3, 4, 5), True)
-        self.assertEqual(config.settings.value(config.Key.Fiducial_Colour), item.default)
+        item = config.settings.default(config.settings.Key.Fiducial_Colour)
+        self.assertEqual(config.settings.value(config.settings.Key.Fiducial_Colour), item.default)
+        config.settings.setValue(config.settings.Key.Fiducial_Colour, (2, 3, 4, 5), True)
+        self.assertEqual(config.settings.value(config.settings.Key.Fiducial_Colour), item.default)
+        config.settings.setValue(config.settings.Key.Fiducial_Colour, (2, 3, 4), True)
+        self.assertEqual(config.settings.value(config.settings.Key.Fiducial_Colour), item.default)
+        config.settings.setValue(config.settings.Key.Fiducial_Colour, ("h", "1.0", "1.0", "1.0"), True)
+        self.assertEqual(config.settings.value(config.settings.Key.Fiducial_Colour), item.default)
+        config.settings.setValue(config.settings.Key.Fiducial_Colour, ("1.0", "1.0", "1.0", "1.0"), True)
+        self.assertEqual(config.settings.value(config.settings.Key.Fiducial_Colour), (1, 1, 1, 1))
+        config.settings.setValue(config.settings.Key.Fiducial_Colour, (2, 3, 4, 5), True)
+        self.assertEqual(config.settings.value(config.settings.Key.Fiducial_Colour), item.default)
 
-        item = config.__defaults__[config.Key.Geometry]
-        self.assertEqual(config.settings.value(config.Key.Geometry), item.default)
-        config.settings.setValue(config.Key.Geometry, "12345", True)
-        self.assertEqual(config.settings.value(config.Key.Geometry), item.default)
-        config.settings.setValue(config.Key.Geometry, bytearray(b"12345"), True)
-        self.assertEqual(config.settings.value(config.Key.Geometry), bytearray(b"12345"))
+        item = config.settings.default(config.settings.Key.Geometry)
+        self.assertEqual(config.settings.value(config.settings.Key.Geometry), item.default)
+        config.settings.setValue(config.settings.Key.Geometry, "12345", True)
+        self.assertEqual(config.settings.value(config.settings.Key.Geometry), item.default)
+        config.settings.setValue(config.settings.Key.Geometry, bytearray(b"12345"), True)
+        self.assertEqual(config.settings.value(config.settings.Key.Geometry), bytearray(b"12345"))
 
-        item = config.__defaults__[config.Key.Recent_Projects]
-        self.assertEqual(config.settings.value(config.Key.Recent_Projects), item.default)
-        config.settings.setValue(config.Key.Recent_Projects, "name", True)
-        self.assertEqual(config.settings.value(config.Key.Recent_Projects), ["name"])
-        config.settings.setValue(config.Key.Recent_Projects, ["name", "other"], True)
-        self.assertEqual(config.settings.value(config.Key.Recent_Projects), ["name", "other"])
+        item = config.settings.default(config.settings.Key.Recent_Projects)
+        self.assertEqual(config.settings.value(config.settings.Key.Recent_Projects), item.default)
+        config.settings.setValue(config.settings.Key.Recent_Projects, "name", True)
+        self.assertEqual(config.settings.value(config.settings.Key.Recent_Projects), ["name"])
+        config.settings.setValue(config.settings.Key.Recent_Projects, ["name", "other"], True)
+        self.assertEqual(config.settings.value(config.settings.Key.Recent_Projects), ["name", "other"])
 
         config.settings.system.sync()
         self.assertTrue(self.ini_file.samefile(config.settings.filename()))
 
-        config.settings.setValue(config.Key.Align_First, False, True)
+        config.settings.setValue(config.settings.Key.Align_First, False, True)
         config.settings.reset()
-        self.assertFalse(config.settings.value(config.Key.Align_First))
+        self.assertFalse(config.settings.value(config.settings.Key.Align_First))
         config.settings.reset(True)
-        self.assertTrue(config.settings.value(config.Key.Align_First))
-        config.settings.setValue(config.Key.Align_First, False)
-        self.assertNotEqual(config.settings.value(config.Key.Align_First),
-                            config.settings.system.value(config.Key.Align_First.value))
+        self.assertTrue(config.settings.value(config.settings.Key.Align_First))
+        config.settings.setValue(config.settings.Key.Align_First, False)
+        self.assertNotEqual(config.settings.value(config.settings.Key.Align_First),
+                            config.settings.system.value(config.settings.Key.Align_First.value))
 
         self.window.showPreferences()
         preferences = self.window.findChild(Preferences)
